// typescript-version/src/types/core.ts
export interface LngLat {
  /** Longitude in degrees [-180, 180] */
  lng: number;
  /** Latitude in degrees [-90, 90] */
  lat: number;
}